refactor: rename Modal, move types/constants to lib, update components exports/imports
