+++ src/lib/constants.ts
@@ -1,0 +1,11 @@
+export const EXTRA_COLUMNS = [
+  'methane',
+  'methane_per_capita',
+  'nitrous_oxide',
+  'temperature_change_from_co2',
+  'oil_co2',
+  'total_ghg',
+  'land_use_change_co2'
+];
+
+export const LAST_YEAR = 2023;
